refactor: migrate to ES modules and update MongoDB connection handling
