+++ backend/package.json
@@ -1,6 +1,7 @@
 {
   "name": "aurabus-api",
   "version": "1.0.0",
+  "type": "module",
   "main": "src/server.js",
   "scripts": {
     "start": "node src/server.js",
