+++ backend/src/server.js
@@ -1,8 +1,19 @@
-const app = require('./app');
-const config = require('./config');
+import {app, connectDb, config} from './app.js';
 
-const port = config.api.port;
+const port = config.api.port; 
 
-app.listen(port, () => {
-  console.log(`Server API listening on port ${port}`);
-});
+async function startServer() {
+  try {
+    await connectDb();
+
+    app.listen(port, () => {
+      console.log(`Server API listening on port ${port}`);
+    });
+
+  } catch (err) {
+    console.error('Failed to start server:', err);
+    process.exit(1);
+  }
+}
+
+startServer();
